Tweet resolutions
This completes [waffle card 10](https://waffle.io/davemaurer/resolutionary/cards/56045f118a19da3700d45a89)

We have decided to start actually writing tests, so this was the first feature
on the app that was test driven. The helper methods in the tests are a bit
dense, as we needed to jump through some hoops to mock out Twitter, and we
don't yet have enough context to figure out a better way to extract this out
and have it be a bit less noisy. At the very least, the test itself
demonstrates intent pretty well.

The feature was fairly straightforward to implement, and I don't think there
was much that stood out as weird. We learned about keyword arguments, which we
hadn't seen before, and learned how to use. We also stumbled a little bit when
we realized that even when we're making an expectation on a request having been
made, we still need to stub it before the request is actually made, since
WebMock doesn't know to allow it.